+++ examples/nuxt/plugins/walletManager.client.ts
@@ -11,6 +11,10 @@ export default defineNuxtPlugin((nuxtApp) => {
         id: WalletId.WALLETCONNECT,
         options: { projectId: 'fcfde0713d43baa0d23be0773c80a72b' }
       },
+      {
+        id: WalletId.BIATEC,
+        options: { projectId: 'fcfde0713d43baa0d23be0773c80a72b' }
+      },
       WalletId.KMD,
       WalletId.KIBISIS,
       {