+++ examples/vanilla-ts/src/main.ts
@@ -14,6 +14,10 @@ const walletManager = new WalletManager({
       id: WalletId.WALLETCONNECT,
       options: { projectId: 'fcfde0713d43baa0d23be0773c80a72b' }
     },
+    {
+      id: WalletId.BIATEC,
+      options: { projectId: 'fcfde0713d43baa0d23be0773c80a72b' }
+    },
     WalletId.KMD,
     WalletId.KIBISIS,
     {