+++ packages/use-wallet/src/wallets/kibisis.ts
@@ -105,10 +105,11 @@ export class KibisisWallet extends BaseWallet {
           )
         }
 
+        this.logger.debug('Disable successful', { result })
         return resolve(result)
       })
 
-      // send the request
+      this.logger.debug('Sending disable request...', { genesisHash })
       avmWebClient.disable({
         genesisHash,
         providerId: KIBISIS_AVM_WEB_PROVIDER_ID
@@ -169,10 +170,11 @@ export class KibisisWallet extends BaseWallet {
           )
         }
 
+        this.logger.debug('Enable successful', { result })
         return resolve(result)
       })
 
-      // send the request
+      this.logger.debug('Sending enable request...', { genesisHash })
       avmWebClient.enable({
         genesisHash,
         providerId: KIBISIS_AVM_WEB_PROVIDER_ID
@@ -188,33 +190,35 @@ export class KibisisWallet extends BaseWallet {
   }
 
   private async _initializeAVMWebClient(): Promise<AVMWebProviderSDK.AVMWebClient> {
-    const _functionName = '_initializeAVMWebClient'
     const avmWebProviderSDK = this.avmWebProviderSDK || (await this._initializeAVMWebProviderSDK())
 
     if (!this.avmWebClient) {
-      console.info(`[${KibisisWallet.name}]#${_functionName}: initializing new client...`)
-
+      this.logger.info('Initializing new AVM Web Client...')
       this.avmWebClient = avmWebProviderSDK.AVMWebClient.init()
+      this.logger.info('AVM Web Client initialized')
     }
 
     return this.avmWebClient
   }
 
   private async _initializeAVMWebProviderSDK(): Promise<typeof AVMWebProviderSDK> {
-    const _functionName = '_initializeAVMWebProviderSDK'
-
     if (!this.avmWebProviderSDK) {
-      console.info(
-        `[${KibisisWallet.name}]#${_functionName}: initializing @agoralabs-sh/avm-web-provider...`
-      )
-
-      this.avmWebProviderSDK = await import('@agoralabs-sh/avm-web-provider')
+      this.logger.info('Initializing @agoralabs-sh/avm-web-provider...')
+      const module = await import('@agoralabs-sh/avm-web-provider')
+      this.avmWebProviderSDK = module.default ? module.default : module
 
       if (!this.avmWebProviderSDK) {
         throw new Error(
-          'failed to initialize, the @agoralabs-sh/avm-web-provider sdk was not provided'
+          'Failed to initialize, the @agoralabs-sh/avm-web-provider SDK was not provided'
         )
       }
+
+      if (!this.avmWebProviderSDK.AVMWebClient) {
+        throw new Error(
+          'Failed to initialize, AVMWebClient missing from @agoralabs-sh/avm-web-provider SDK'
+        )
+      }
+      this.logger.info('@agoralabs-sh/avm-web-provider initialized')
     }
 
     return this.avmWebProviderSDK
@@ -287,10 +291,11 @@ export class KibisisWallet extends BaseWallet {
           )
         }
 
+        this.logger.debug('Sign transactions successful', { result })
         return resolve(result)
       })
 
-      // send the request
+      this.logger.debug('Sending sign transactions request...', { txns })
       avmWebClient.signTransactions({
         txns,
         providerId: KIBISIS_AVM_WEB_PROVIDER_ID
@@ -306,19 +311,13 @@ export class KibisisWallet extends BaseWallet {
     let result: AVMWebProviderSDK.IEnableResult
 
     try {
-      console.info(`[${this.metadata.name}] Connecting...`)
-
+      this.logger.info('Connecting...')
       result = await this._enable()
-
-      console.info(
-        `[${this.metadata.name}] Successfully connected on network "${result.genesisId}"`
-      )
+      this.logger.info(`Successfully connected on network "${result.genesisId}"`)
     } catch (error: any) {
-      console.error(
-        `[${this.metadata.name}] Error connecting: ` +
-          (isAVMWebProviderSDKError(error)
-            ? `${error.message} (code: ${error.code})`
-            : error.message)
+      this.logger.error(
+        'Error connecting:',
+        isAVMWebProviderSDKError(error) ? `${error.message} (code: ${error.code})` : error.message
       )
       throw error
     }
@@ -335,26 +334,24 @@ export class KibisisWallet extends BaseWallet {
       wallet: walletState
     })
 
-    console.info(`[${this.metadata.name}] ✅ Connected.`, walletState)
+    this.logger.info('✅ Connected.', walletState)
     return walletAccounts
   }
 
   public async disconnect(): Promise<void> {
     try {
-      console.info(`[${this.metadata.name}] Disconnecting...`)
+      this.logger.info('Disconnecting...')
       this.onDisconnect()
 
       const result = await this._disable()
 
-      console.info(
-        `[${this.metadata.name}] Successfully disconnected${result.sessionIds && result.sessionIds.length ? ` sessions [${result.sessionIds.join(',')}]` : ''} on network "${result.genesisId}"`
+      this.logger.info(
+        `Successfully disconnected${result.sessionIds && result.sessionIds.length ? ` sessions [${result.sessionIds.join(',')}]` : ''} on network "${result.genesisId}"`
       )
     } catch (error: any) {
-      console.error(
-        `[${this.metadata.name}] Error disconnecting: ` +
-          (isAVMWebProviderSDKError(error)
-            ? `${error.message} (code: ${error.code})`
-            : error.message)
+      this.logger.error(
+        'Error disconnecting:',
+        isAVMWebProviderSDKError(error) ? `${error.message} (code: ${error.code})` : error.message
       )
       throw error
     }
@@ -366,23 +363,24 @@ export class KibisisWallet extends BaseWallet {
     let result: AVMWebProviderSDK.IEnableResult
 
     if (!walletState) {
+      this.logger.info('No session to resume')
       return
     }
 
     try {
-      console.info(`[${this.metadata.name}] Resuming session...`)
+      this.logger.info('Resuming session...')
 
       result = await this._enable()
 
       if (result.accounts.length === 0) {
-        throw new Error(`No accounts found!`)
+        throw new Error('No accounts found!')
       }
 
       const walletAccounts = this._mapAVMWebProviderAccountToWalletAccounts(result.accounts)
       const match = compareAccounts(walletAccounts, walletState.accounts)
 
       if (!match) {
-        console.warn(`[${this.metadata.name}] Session accounts mismatch, updating accounts`, {
+        this.logger.warn('Session accounts mismatch, updating accounts', {
           prev: walletState.accounts,
           current: walletAccounts
         })
@@ -392,12 +390,11 @@ export class KibisisWallet extends BaseWallet {
           accounts: walletAccounts
         })
       }
+      this.logger.info('Session resumed successfully')
     } catch (error: any) {
-      console.error(
-        `[${this.metadata.name}] Error resuming session: ` +
-          (isAVMWebProviderSDKError(error)
-            ? `${error.message} (code: ${error.code})`
-            : error.message)
+      this.logger.error(
+        'Error resuming session:',
+        isAVMWebProviderSDKError(error) ? `${error.message} (code: ${error.code})` : error.message
       )
       this.onDisconnect()
       throw error
@@ -465,6 +462,7 @@ export class KibisisWallet extends BaseWallet {
     indexesToSign?: number[]
   ): Promise<(Uint8Array | null)[]> => {
     try {
+      this.logger.debug('Signing transactions...', { txnGroup, indexesToSign })
       let txnsToSign: AVMWebProviderSDK.IARC0001Transaction[] = []
 
       // Determine type and process transactions for signing
@@ -476,8 +474,11 @@ export class KibisisWallet extends BaseWallet {
         txnsToSign = this.processEncodedTxns(flatTxns, indexesToSign)
       }
 
+      this.logger.debug('Sending processed transactions to wallet...', txnsToSign)
+
       // Sign transactions
       const signTxnsResult = await this._signTransactions(txnsToSign)
+      this.logger.debug('Received signed transactions from wallet', signTxnsResult)
 
       // Convert base64 to Uint8Array
       const result = signTxnsResult.stxns.map((value) => {
@@ -487,13 +488,12 @@ export class KibisisWallet extends BaseWallet {
         return base64ToByteArray(value)
       })
 
+      this.logger.debug('Transactions signed successfully', result)
       return result
     } catch (error: any) {
-      console.error(
-        `[${this.metadata.name}] error signing transactions: ` +
-          (isAVMWebProviderSDKError(error)
-            ? `${error.message} (code: ${error.code})`
-            : error.message)
+      this.logger.error(
+        'Error signing transactions:',
+        isAVMWebProviderSDKError(error) ? `${error.message} (code: ${error.code})` : error.message
       )
       throw error
     }